src/pyproject.toml
[tool.poetry]
name = "trip-estimator"
version = "0.1.0"
description = ""
authors = ["TheLastAurora <vitoribeiro0024@gmail.com>"]
packages = [{include = "services", from="src"}]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio-tools = "^1.64.1"
grpcio = "^1.64.1"
protobuf = "^5.27.2"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
